backend/requirements.txt

python-dotenv
flask
pymongo
dnspython
flask-cors
pycryptodome
psutil
ecdsa>=0.18.0
gunicorn


